Rename(gameResultService): service 패키지 하위로 이동, 이름 변경
