+++ src/test/java/baseball/BaseballGameTest.java
@@ -1,7 +1,7 @@
 package baseball;
 
 import baseball.model.ComputerNumber;
-import baseball.model.GameResultGenerator;
+import baseball.service.GameResultService;
 import baseball.model.PlayerChoice;
 import baseball.model.PlayerNumber;
 import camp.nextstep.edu.missionutils.Randoms;
@@ -53,10 +53,10 @@ public class BaseballGameTest {
 
             ComputerNumber computerNumber = ComputerNumber.create();  //1,2,3
             PlayerNumber playerNumber = PlayerNumber.create("456");
-            GameResultGenerator gameResultGenerator = new GameResultGenerator(computerNumber, playerNumber);
-            assertThat(gameResultGenerator.isNothing()).isTrue();
-            assertThat(gameResultGenerator.getBallCount()).isEqualTo(0);
-            assertThat(gameResultGenerator.getStrikeCount()).isEqualTo(0);
+            GameResultService gameResultService = new GameResultService(computerNumber, playerNumber);
+            assertThat(gameResultService.isNothing()).isTrue();
+            assertThat(gameResultService.getBallCount()).isEqualTo(0);
+            assertThat(gameResultService.getStrikeCount()).isEqualTo(0);
         }
     }
 
@@ -69,10 +69,10 @@ public class BaseballGameTest {
 
             ComputerNumber computerNumber = ComputerNumber.create();  //1,2,3
             PlayerNumber playerNumber = PlayerNumber.create("451");  //4,5,1
-            GameResultGenerator gameResultGenerator = new GameResultGenerator(computerNumber, playerNumber);
-            assertThat(gameResultGenerator.isNothing()).isFalse();
-            assertThat(gameResultGenerator.getBallCount()).isEqualTo(1);
-            assertThat(gameResultGenerator.getStrikeCount()).isEqualTo(0);
+            GameResultService gameResultService = new GameResultService(computerNumber, playerNumber);
+            assertThat(gameResultService.isNothing()).isFalse();
+            assertThat(gameResultService.getBallCount()).isEqualTo(1);
+            assertThat(gameResultService.getStrikeCount()).isEqualTo(0);
         }
     }
 
@@ -85,10 +85,10 @@ public class BaseballGameTest {
 
             ComputerNumber computerNumber = ComputerNumber.create();  //1,2,3
             PlayerNumber playerNumber = PlayerNumber.create("351");  //3,5,1
-            GameResultGenerator gameResultGenerator = new GameResultGenerator(computerNumber, playerNumber);
-            assertThat(gameResultGenerator.isNothing()).isFalse();
-            assertThat(gameResultGenerator.getBallCount()).isEqualTo(2);
-            assertThat(gameResultGenerator.getStrikeCount()).isEqualTo(0);
+            GameResultService gameResultService = new GameResultService(computerNumber, playerNumber);
+            assertThat(gameResultService.isNothing()).isFalse();
+            assertThat(gameResultService.getBallCount()).isEqualTo(2);
+            assertThat(gameResultService.getStrikeCount()).isEqualTo(0);
         }
     }
 
@@ -101,10 +101,10 @@ public class BaseballGameTest {
 
             ComputerNumber computerNumber = ComputerNumber.create();  //1,2,3
             PlayerNumber playerNumber = PlayerNumber.create("231");  //2,3,1
-            GameResultGenerator gameResultGenerator = new GameResultGenerator(computerNumber, playerNumber);
-            assertThat(gameResultGenerator.isNothing()).isFalse();
-            assertThat(gameResultGenerator.getBallCount()).isEqualTo(3);
-            assertThat(gameResultGenerator.getStrikeCount()).isEqualTo(0);
+            GameResultService gameResultService = new GameResultService(computerNumber, playerNumber);
+            assertThat(gameResultService.isNothing()).isFalse();
+            assertThat(gameResultService.getBallCount()).isEqualTo(3);
+            assertThat(gameResultService.getStrikeCount()).isEqualTo(0);
         }
     }
 
@@ -117,10 +117,10 @@ public class BaseballGameTest {
 
             ComputerNumber computerNumber = ComputerNumber.create();  //1,2,3
             PlayerNumber playerNumber = PlayerNumber.create("145");  //1,4,5
-            GameResultGenerator gameResultGenerator = new GameResultGenerator(computerNumber, playerNumber);
-            assertThat(gameResultGenerator.isNothing()).isFalse();
-            assertThat(gameResultGenerator.getBallCount()).isEqualTo(0);
-            assertThat(gameResultGenerator.getStrikeCount()).isEqualTo(1);
+            GameResultService gameResultService = new GameResultService(computerNumber, playerNumber);
+            assertThat(gameResultService.isNothing()).isFalse();
+            assertThat(gameResultService.getBallCount()).isEqualTo(0);
+            assertThat(gameResultService.getStrikeCount()).isEqualTo(1);
         }
     }
 
@@ -133,10 +133,10 @@ public class BaseballGameTest {
 
             ComputerNumber computerNumber = ComputerNumber.create();  //1,2,3
             PlayerNumber playerNumber = PlayerNumber.create("125");  //1,2,5
-            GameResultGenerator gameResultGenerator = new GameResultGenerator(computerNumber, playerNumber);
-            assertThat(gameResultGenerator.isNothing()).isFalse();
-            assertThat(gameResultGenerator.getBallCount()).isEqualTo(0);
-            assertThat(gameResultGenerator.getStrikeCount()).isEqualTo(2);
+            GameResultService gameResultService = new GameResultService(computerNumber, playerNumber);
+            assertThat(gameResultService.isNothing()).isFalse();
+            assertThat(gameResultService.getBallCount()).isEqualTo(0);
+            assertThat(gameResultService.getStrikeCount()).isEqualTo(2);
         }
     }
 
@@ -149,10 +149,10 @@ public class BaseballGameTest {
 
             ComputerNumber computerNumber = ComputerNumber.create();  //1,2,3
             PlayerNumber playerNumber = PlayerNumber.create("123");  //1,4,5
-            GameResultGenerator gameResultGenerator = new GameResultGenerator(computerNumber, playerNumber);
-            assertThat(gameResultGenerator.isNothing()).isFalse();
-            assertThat(gameResultGenerator.getBallCount()).isEqualTo(0);
-            assertThat(gameResultGenerator.getStrikeCount()).isEqualTo(3);
+            GameResultService gameResultService = new GameResultService(computerNumber, playerNumber);
+            assertThat(gameResultService.isNothing()).isFalse();
+            assertThat(gameResultService.getBallCount()).isEqualTo(0);
+            assertThat(gameResultService.getStrikeCount()).isEqualTo(3);
         }
     }
 
@@ -165,10 +165,10 @@ public class BaseballGameTest {
 
             ComputerNumber computerNumber = ComputerNumber.create();  //1,2,3
             PlayerNumber playerNumber = PlayerNumber.create("142");  //1,4,2
-            GameResultGenerator gameResultGenerator = new GameResultGenerator(computerNumber, playerNumber);
-            assertThat(gameResultGenerator.isNothing()).isFalse();
-            assertThat(gameResultGenerator.getBallCount()).isEqualTo(1);
-            assertThat(gameResultGenerator.getStrikeCount()).isEqualTo(1);
+            GameResultService gameResultService = new GameResultService(computerNumber, playerNumber);
+            assertThat(gameResultService.isNothing()).isFalse();
+            assertThat(gameResultService.getBallCount()).isEqualTo(1);
+            assertThat(gameResultService.getStrikeCount()).isEqualTo(1);
         }
     }
 
@@ -181,10 +181,10 @@ public class BaseballGameTest {
 
             ComputerNumber computerNumber = ComputerNumber.create();  //1,2,3
             PlayerNumber playerNumber = PlayerNumber.create("132");  //1,3,2
-            GameResultGenerator gameResultGenerator = new GameResultGenerator(computerNumber, playerNumber);
-            assertThat(gameResultGenerator.isNothing()).isFalse();
-            assertThat(gameResultGenerator.getBallCount()).isEqualTo(2);
-            assertThat(gameResultGenerator.getStrikeCount()).isEqualTo(1);
+            GameResultService gameResultService = new GameResultService(computerNumber, playerNumber);
+            assertThat(gameResultService.isNothing()).isFalse();
+            assertThat(gameResultService.getBallCount()).isEqualTo(2);
+            assertThat(gameResultService.getStrikeCount()).isEqualTo(1);
         }
     }
